Add initial version of auto activation code
Using Python's virtual environments for isolation from other packages
installed on the system can be cumbersome and activation is easy to
forget. To that end, an automated approach is desired.
This change introduces pyvenv-auto, a shell script that handles
activation and deactivation automatically. The idea is to overwrite the
'cd' built-in with a custom function that checks for virtual env
structure (that is, a 'pyvenv.cfg' configuration file in the directory
being changed to and a 'activate' script in a bin/ sub directory). The
script will then source the 'activate' script in order to activate the
virtual environment.
The script is designed to keep overhead low. For that matter, it only
performs checks on every 'cd' and not everytime a command is executed
(as would be the case with a PROMPT_COMMAND based solution). The
rationale here is that virtual envs do not change frequently. If a
virtual environment is created in the current directory it will not be
activated immediately automatically but activation can be forced by
executing a 'cd .'.
Setup of the script itself is simple, one mainly has to source it in a
shell initialization file (e.g., ~/.bashrc) and have it evaluate the
current directory, like so:
  source ~/.pyvenv-auto.sh
  _pyvenv_activate_deactivate