Only check for existence of files required for pyvenv and not type
In order to detect whether or not a directory has the proper structure
of a virtual environment (that is, it contains a pyvenv.cfg
configuration and a bin/activate file), we use 'test -f'. This test
implies that the files are regular files. However, there are reasons for
making them symlinks in which case a virtual environment would not be
detected.
To that end, this change loosens the check even more by purely testing
for existence by means of 'test -e'.