Add section about potential prompt clashes to README
Virtual environments typically update the shell prompt in use to
indicate that a particular environment is active. This update can clash
with updates made for other reasons, to indicate the status of a git
repository, for example.
As this is a potentially common problem with an existing solution, this
change dedicates a section in the README file to explaining the problem
and proposing a solution.