Invoke deactivate with an argument
With strict variable reference checking enabled, we see errors being
reported when deactivating a virtual environment by changing out of its
containing directory. The problem is that the deactivate function as
provided by the virtual environment unconditionally references ${1}
which does not exist without such an argument.
To work around this problem, this patch invokes the deactivate function
with a single empty string being supplied as an argument.